Module differential_dataflow::collection

A time-varying multiset of records.

The core datastructure representing a dataset in differential dataflow is collection::Trace. It represents a multiset that can be indexed by partially ordered times, and is stored in a compressed incremental representation.

Reexports

pub use collection::lookup::Lookup;

pub use collection::least_upper_bound::LeastUpperBound;

pub use collection::least_upper_bound::close_under_lub;

pub use collection::trace::Trace;

pub use collection::trace::Offset;

Modules

compact
Accumulator of (key, val, wgt) triples based on sorting and run-length encoding.

count
Like Count but with the value type specialized to ().

least_upper_bound

lookup

robin_hood

tier
A trie representation of (key, time, value, weight) tuples, and routines to merge them.

trace