Struct differential_dataflow::collection::compact::Compact

pub struct Compact<K, V> {
    pub keys: Vec<K>,
    pub cnts: Vec<u32>,
    pub vals: Vec<(V, i32)>,
}

A compressed representation of the accumulation of (key, val, wgt) triples.

Fields

keys

An ordered list of the distinct keys.

cnts

Counts for each key indicating the number of corresponding values in self.vals.

The list is maintained separately in the interest of eventually having run-length coding treat non-repetitions better.

vals

A list of values, ordered within each key group.

Methods

impl<K: Ord + Debug, V: Ord> Compact<K, V>

fn new(k: usize, v: usize) -> Compact<K, V>

Constructs a new Compact with indicated initial capacities.

Most operations with Compact eventually shrink the amount of memory to fit whatever they have used, so the main concern here is to avoid grossly over-allocating. Typically, these structs are created in a transient compaction step and not maintained open, meaning we can afford to be a bit sloppy.

fn size(&self) -> usize

Reports the size in bytes, used elsewhere to determine how much space we should use for buffering uncompressed elements.

fn extend<I: Iterator<Item=((K, V), i32)>>(&mut self, iterator: I)

Populates the Compact from an iterator of ordered (key, val, wgt) triples.

The Compact does not know about the ordering, only that it should look for repetitions of in the sequences of key and wgt.

fn extend_by(&mut self, buffer: &mut Vec<((K, V), i32)>)

fn from_radix<U: Unsigned + Default, F: Fn(&K) -> U>(source: &mut Vec<Vec<((K, V), i32)>>, function: &F) -> Option<Compact<K, V>>

fn session<'a>(&'a mut self) -> CompactSession<'a, K, V>

fn push<I: Iterator<Item=(V, i32)>>(&mut self, key: K, iterator: I)