Struct differential_dataflow::collection::count::Count

pub struct Count<K, T, L> {
    // some fields omitted
}

A trace specialized for the value type ().

This trace does not need to represent a variable number of values associated with each key, and instead needs only one weight per timestamp.

Methods

impl<K, L, T> Count<K, T, L> where K: Data + Ord + 'static, L: Lookup<K, Offset> + 'static, T: Lattice + 'static

fn get_count(&self, key: &K, time: &T) -> i32

Recovers the count for the () value for a key at a time.

impl<K: Eq, L: Lookup<K, Offset>, T> Count<K, T, L>

fn new(l: L) -> Count<K, T, L>

Allocates a new Count from a supplied lookup.