Module differential_dataflow::collection::robin_hood

Structs

RHHMap

Functions

get_mut
Returns either an exclusive reference to the value associated with query, or None if no value exists.

get_ref
Returns either a shared reference to the value associated with query, or None if no value exists.

insert
Returns either the previous (key,val) pair for key, or Err((key,val)) if slice does not have enough room to insert it.

remove
Returns either the (key,val) pair associated with key, or None if it does not exist.