Struct differential_dataflow::collection::tier::Merge

pub struct Merge<K, T, V> {
    // some fields omitted
}

A merge-in-progress of two instances of Tier<K, T, V>.

A Merge represents a partial merge of two instances of Tier<K, T, V> into one instance, where times are advanced according to a function advance supplied to the new constructor, and like (K, T, V, _) tuples are consolidated into at most on result tuple.

A Merge can execute progressively, allowing a large amount of work to be amortized over the large number of tuples involved. The MergePart structs contained in a Merge use Rc<Tier> fields to capture their tiers, as the Merge does not mutate the source Tier instances.

Methods

impl<K: Ord + Clone, T: Eq + Clone + Hash, V: Ord + Clone> Merge<K, T, V>

fn new<F: Fn(&T) -> T>(tier1: &Rc<Tier<K, T, V>>, tier2: &Rc<Tier<K, T, V>>, advance: &F) -> Merge<K, T, V>

Constructs a new Merge from two instances of Teir and a function advancing timestamps.

This method initiates a merge of two tiers, consolidating their representation which can potentially reduce the complexity and amount of memory required to describe a trace. The required inputs are two Rc<Tier<K, T, V>> instances to merge, and a function advance from &T to T indicating how timestamps should be advanced.

As part of initiating the merge, new will scan through the timestamps used by each source Tier, advancing each and creating a mapping from timestamp indices in each source to new advanced and unified timestamp indices.

fn step(&mut self) -> Option<Tier<K, T, V>>

Advances the Merge by one step, returning the merged tiers if it is now complete.

The step method considers the next key proposed by tier1 and tier2, and populates self.result as appropriate. If both tier1 and tier2 have the same key and the merged results cancel one-another, self.result may not actually change (although the merge has performed useful work).

The step method returns Some(result) if the merge is now complete, and None if it is not yet complete, and should be called more.

Once step returns a result, the Merge contains an empty result field. It is then a logic error to do anything other than discard the Merge, though the usage patterns don't currently enforce this.