Struct differential_dataflow::collection::tier::Tier

pub struct Tier<K, T, V> {
    pub keys: Vec<(K, usize)>,
    pub idxs: Vec<(usize, usize)>,
    pub vals: Vec<(V, W)>,
    pub times: Vec<(T, usize)>,
}

A collection of (K, T, V, W) tuples, grouped by K then T then V.

A Tier is a trie-representation of (K, T, V, W) tuples, meaning its representation is as a sorted list of these tuples, where each of the fields are then stored separately and run-length coded. The keys field is a sorted list of pairs (K, usize), indicating a key and an offset in the next-level array, idxs. Likewise, the idxs field contains a flat list of pairs (usize, usize), where each interval described by one entry of keys is sorted. Each entry (usize, usize) indicates a time (element of the times field), and an offset in the vals field. Finally, the vals field has each interval sorted by V.

Fields

keys

Pairs of key and offset into self.idxs.

idxs

Pairs of idx and offset into self.vals.

vals

Pairs of val and weight.

times

Pairs of timestamp and the number of references to the timestamp.

Methods

impl<K: Ord, T: Eq, V: Ord + Clone> Tier<K, T, V>

fn new() -> Tier<K, T, V>

Constructs a new Tier containing no data.

fn with_capacities(k: usize, i: usize, v: usize) -> Tier<K, T, V>

Allocates a new Tier with initial capacities for keys, idxs, and vals.

fn len(&self) -> usize

Returns the number of tuples represented by Self.