Struct differential_dataflow::collection::trace::Trace

pub struct Trace<K, T, V, L> {
    pub keys: L,
    // some fields omitted
}

A collection of values indexed by key and time.

Safety

For reasons of borrow checking, it is difficult to merge references using already-allocated memory. The method get_collection_using uses its heap argument to store references that should remain valid for as long as the Trace is valid, but I cannot convince Rust of this fact because only I know that once installed, differences are immutable. Please do not call get_collection_using with a heap argument that may out-live the Trace itself.

Fields

keys

Methods

impl<K, V, L, T> Trace<K, T, V, L> where K: Ord, V: Ord, L: Lookup<K, Offset>, T: LeastUpperBound + Debug

fn set_difference(&mut self, time: T, accumulation: Compact<K, V>)

Installs a supplied set of keys and values as the differences for time.

fn get_difference<'a>(&'a self, key: &K, time: &T) -> DifferenceIterator<'a, V>

Enumerates the differences for key at time.

unsafe fn get_collection_using<'a>(&'a self, key: &K, time: &T, heap: &mut Vec<((&(), i32), DifferenceIterator<'static, ()>)>) -> CollectionIterator<'a, V>

Enumerates the collection for key at time.

A collection is defined as the accumulation of all differences at times less or equal to time.

fn interesting_times<'a>(&'a mut self, key: &K, index: T) -> &'a [T]

Lists times that are the least upper bound of time and any subset of existing times.

fn trace<'a>(&'a self, key: &K) -> TraceIterator<'a, K, T, V, L>

Enumerates pairs of time &T and differences DifferenceIterator<V> for key.

impl<K: Eq, L: Lookup<K, Offset>, T, V> Trace<K, T, V, L>

fn new(l: L) -> Trace<K, T, V, L>