Module differential_dataflow::difference

A type that can be treated as a difference.

Differential dataflow most commonly tracks the counts associated with records in a multiset, but it generalizes to tracking any map from the records to an Abelian group. The most common generalization is when we maintain both a count and another accumulation, for example height. The differential dataflow collections would then track for each record the total of counts and heights, which allows us to track something like the average.

Structs

DiffPair
The difference defined by a pair of difference elements.

Traits

Diff
A type that can be treated as a difference.