Trait differential_dataflow::lattice::TotalOrder

pub trait TotalOrder: Lattice { }

A carrier trait for totally ordered lattices.

Types that implement TotalOrder are stating that their Lattice is in fact a total order. This type is only used to restrict implementations to certain types of lattices.

This trait is automatically implemented for integer scalars, and for products of these types with "empty" timestamps (e.g. RootTimestamp and ()). Be careful implementing this trait for your own timestamp types, as it may lead to the applicability of incorrect implementations.

Note that this trait is distinct from Ord; many implementors of Lattice also implement Ord so that they may be sorted, deduplicated, etc. This implementation neither derives any information from an Ord implementation nor informs it in any way.

Examples

use differential_dataflow::lattice::TotalOrder;

// The `join` and `meet` of totally ordered elements are always one of the two.
fn invariant<T: TotalOrder>(elt1: T, elt2: T) {
    if elt1.less_equal(&elt2) {
        assert!(elt1.meet(&elt2) == elt1);
        assert!(elt1.join(&elt2) == elt2);
    }
    else {
        assert!(elt1.meet(&elt2) == elt2);
        assert!(elt1.join(&elt2) == elt1);
    }
}