Module differential_dataflow::operators

Specialize differential dataflow operators.

Differential dataflow introduces a small number of specialized operators on collections. These operators have specialized implementations to make them work efficiently, and are in addition to several operations defined directly on the Collection type (e.g. map and filter).

Reexports

pub use self::group::Group;

pub use self::group::Distinct;

pub use self::group::Count;

pub use self::group::consolidate_from;

pub use self::consolidate::Consolidate;

pub use self::iterate::Iterate;

pub use self::join::Join;

pub use self::join::JoinUnsigned;

pub use self::join::JoinCore;

pub use self::count::CountTotal;

Modules

arrange
Arranges a collection into a re-usable trace structure.

consolidate
Aggregates the weights of equal records into at most one record.

count
Group records by a key, and apply a reduction function.

group
Group records by a key, and apply a reduction function.

iterate
Iterative application of a differential dataflow fragment.

join
Match pairs of records based on a key.