Trait differential_dataflow::operators::arrange::Arrange

pub trait Arrange<G: Scope, K, V, R: Diff> where G::Timestamp: Lattice {
    fn arrange<T>(&self,
                  empty_trace: T)
                  -> Arranged<G, K, V, R, TraceAgent<K, V, G::Timestamp, R, T>> where T: Trace<K, V, G::Timestamp, R> + 'static,
              T::Batch: Batch<K, V, G::Timestamp, R>;
}

Arranges something as (Key,Val) pairs according to a type T of trace.

Required Methods

fn arrange<T>(&self,
empty_trace: T)
-> Arranged<G, K, V, R, TraceAgent<K, V, G::Timestamp, R, T>> where T: Trace<K, V, G::Timestamp, R> + 'static,
T::Batch: Batch<K, V, G::Timestamp, R>

Arranges a stream of (Key, Val) updates by Key. Accepts an empty instance of the trace type.

This operator arranges a stream of values into a shared trace, whose contents it maintains. This trace is current for all times marked completed in the output stream, and probing this stream is the correct way to determine that times in the shared trace are committed.