Trait differential_dataflow::operators::group::GroupBy

pub trait GroupBy<G: Scope, D1: Data>: GroupByCore<G, D1> where G::Timestamp: LeastUpperBound {
    fn group_by_u<U: Data + Unsigned + Default, V1: Data, V2: Data, D2: Data, KV: Fn(D1) -> (U, V1) + 'static, Logic: Fn(&U, &mut CollectionIterator<V1>, &mut Vec<(V2, i32)>) + 'static, Reduc: Fn(&U, &V2) -> D2 + 'static>(&self, kv: KV, reduc: Reduc, logic: Logic) -> Collection<G, D2>;

    fn group_by<K: Data, V1: Data, V2: Data, D2: Data, KV: Fn(D1) -> (K, V1) + 'static, Part: Fn(&D1) -> u64 + 'static, U: Unsigned + Default, KH: Fn(&K) -> U + 'static, Logic: Fn(&K, &mut CollectionIterator<V1>, &mut Vec<(V2, i32)>) + 'static, Reduc: Fn(&K, &V2) -> D2 + 'static>(&self, kv: KV, part: Part, key_h: KH, reduc: Reduc, logic: Logic) -> Collection<G, D2> { ... }
}

Extension trait for the group_by and group_by_u differential dataflow methods.

Required Methods

fn group_by_u<U: Data + Unsigned + Default, V1: Data, V2: Data, D2: Data, KV: Fn(D1) -> (U, V1) + 'static, Logic: Fn(&U, &mut CollectionIterator<V1>, &mut Vec<(V2, i32)>) + 'static, Reduc: Fn(&U, &V2) -> D2 + 'static>(&self, kv: KV, reduc: Reduc, logic: Logic) -> Collection<G, D2>

A specialization of the group_by method to the case that the key type K is an unsigned integer, and the strategy for indexing by key is simply to index into a vector.

Provided Methods

fn group_by<K: Data, V1: Data, V2: Data, D2: Data, KV: Fn(D1) -> (K, V1) + 'static, Part: Fn(&D1) -> u64 + 'static, U: Unsigned + Default, KH: Fn(&K) -> U + 'static, Logic: Fn(&K, &mut CollectionIterator<V1>, &mut Vec<(V2, i32)>) + 'static, Reduc: Fn(&K, &V2) -> D2 + 'static>(&self, kv: KV, part: Part, key_h: KH, reduc: Reduc, logic: Logic) -> Collection<G, D2>

Groups input records together by key and applies a reduction function.

group_by transforms a stream of records of type D1 into a stream of records of type D2, by first transforming each input record into a (key, val): (K, V1) pair. For each key with some values, logic is invoked on the key and an value enumerator which presents (V1, i32) pairs, indicating for each value its multiplicity. logic is expected to populate its third argument, a &mut Vec<(V2, i32)> indicating multiplicities of output records. Finally, for each (key,val) : (K,V2) pair produced, reduc is applied to produce an output D2 record.

This all may seem overcomplicated, and it may indeed become simpler in the future. For the moment it is designed to allow as much programmability as possible.