Struct differential_dataflow::operators::iterate::CoreVariable

pub struct CoreVariable<'a, G: Scope, D: Data, R: Diff, T: Timestamp + Lattice> where
    G::Timestamp: Lattice,  { /* fields omitted */ }

A recursive variable with a generic timestamp type.

This type generalizes Variable by permitting an arbitrary timestamp type, but incurs the additional cost of validating the step of the timestamp, which Variable does not do. This is technically a bug in Variable, as supplying a timestamp that overflows (using e.g. enter_at) would result in errors.

Methods

impl<'a, G: Scope, D: Data, R: Diff, T: Timestamp + Lattice> CoreVariable<'a, G, D, R, T> where
G::Timestamp: Lattice,

pub fn from_args(
max_steps: T,
step: <T as Timestamp>::Summary,
source: Collection<Child<'a, G, T>, D, R>
) -> CoreVariable<'a, G, D, R, T>

Creates a new Variable and a Stream representing its output, from a supplied source stream.

pub fn set(
self,
result: &Collection<Child<'a, G, T>, D, R>
) -> Collection<Child<'a, G, T>, D, R>

Adds a new source of data to the Variable.