Struct differential_dataflow::operators::iterate::Variable

pub struct Variable<'a, G: Scope, D: Data, R: Diff> where G::Timestamp: Lattice { /* fields omitted */ }

A differential dataflow collection variable

The Variable struct allows differential dataflow programs requiring more sophisticated iterative patterns than singly recursive iteration. For example: in mutual recursion two collections evolve simultaneously.

Examples

The following example is equivalent to the example for the Iterate trait.

extern crate timely;
extern crate differential_dataflow;

use timely::dataflow::Scope;

use differential_dataflow::input::Input;
use differential_dataflow::operators::iterate::Variable;
use differential_dataflow::operators::Consolidate;

fn main() {
    ::timely::example(|scope| {

        let numbers = scope.new_collection_from(1 .. 10u32).1;

        scope.scoped(|nested| {
            let variable = Variable::from(numbers.enter(nested));
            let result = variable.map(|x| if x % 2 == 0 { x/2 } else { x })
                                 .consolidate();
            variable.set(&result)
                    .leave()
        });
    })
}

Methods

impl<'a, G: Scope, D: Data, R: Diff> Variable<'a, G, D, R> where G::Timestamp: Lattice

fn from(source: Collection<Child<'a, G, u64>, D, R>) -> Variable<'a, G, D, R>

Creates a new Variable and a Stream representing its output, from a supplied source stream.

fn set(self,
result: &Collection<Child<'a, G, u64>, D, R>)
-> Collection<Child<'a, G, u64>, D, R>

Adds a new source of data to the Variable.