Trait differential_dataflow::operators::iterate::IterateExt

pub trait IterateExt<G: Scope, D: Data> {
    fn iterate<F>(&self, logic: F) -> Collection<G, D> where G::Timestamp: LeastUpperBound, F: FnOnce(&Collection<Child<G, u64>, D>) -> Collection<Child<G, u64>, D>;
    fn iterate_by<P, U, F>(&self, partition: P, logic: F) -> Collection<G, D> where G::Timestamp: LeastUpperBound, U: Unsigned, P: Fn(&D) -> U + 'static, F: FnOnce(&Collection<Child<G, u64>, D>) -> Collection<Child<G, u64>, D>;
}

An extension trait for the iterate method.

Required Methods

fn iterate<F>(&self, logic: F) -> Collection<G, D> where G::Timestamp: LeastUpperBound, F: FnOnce(&Collection<Child<G, u64>, D>) -> Collection<Child<G, u64>, D>

Iteratively apply logic to the source collection until convergence.

fn iterate_by<P, U, F>(&self, partition: P, logic: F) -> Collection<G, D> where G::Timestamp: LeastUpperBound, U: Unsigned, P: Fn(&D) -> U + 'static, F: FnOnce(&Collection<Child<G, u64>, D>) -> Collection<Child<G, u64>, D>

Iteratively apply logic to the source collection until convergence.

The partition argument is used to partition the data for consolidation.