Trait differential_dataflow::operators::join::JoinBy

pub trait JoinBy<G: Scope, D1: Data + Ord>: Binary<G, (D1, i32)> + Map<G, (D1, i32)> where G::Timestamp: LeastUpperBound + Debug {
    fn join_by_u<U: Unsigned + Data + Default, V1: Data + Default + 'static, V2: Data + Default + 'static, D2: Data, F1: Fn(D1) -> (U, V1) + 'static, F2: Fn(D2) -> (U, V2) + 'static, R: Data, RF: Fn(&U, &V1, &V2) -> R + 'static>(&self, other: &Stream<G, (D2, i32)>, kv1: F1, kv2: F2, result: RF) -> Stream<G, (R, i32)> { ... }
    fn semijoin_by_u<U: Unsigned + Data + Default, V1: Data + Default + 'static, F1: Fn(D1) -> (U, V1) + 'static, RF: Fn(&U, &V1) -> D1 + 'static>(&self, other: &Stream<G, (U, i32)>, kv1: F1, result: RF) -> Stream<G, (D1, i32)> { ... }
    fn join_by<K: Data + 'static, V1: Data + 'static, V2: Data + 'static, D2: Data, F1: Fn(D1) -> (K, V1) + 'static, F2: Fn(D2) -> (K, V2) + 'static, U: Unsigned + Data + Default, KH: Fn(&K) -> U + 'static, R: Data, RF: Fn(&K, &V1, &V2) -> R + 'static>(&self, other: &Stream<G, (D2, i32)>, kv1: F1, kv2: F2, key_h: KH, result: RF) -> Stream<G, (R, i32)> { ... }
    fn semijoin_by<K: Data + 'static, V1: Data + 'static, F1: Fn(D1) -> (K, V1) + 'static, KH: Fn(&K) -> u64 + 'static, RF: Fn(&K, &V1) -> D1 + 'static>(&self, other: &Stream<G, (K, i32)>, kv1: F1, key_h: KH, result: RF) -> Stream<G, (D1, i32)> { ... }
    fn join_by_core<K: Data, V1: Data, V2: Data, D2: Data, F1: Fn(D1) -> (K, V1) + 'static, F2: Fn(D2) -> (K, V2) + 'static, H1: Fn(&D1) -> u64 + 'static, H2: Fn(&D2) -> u64 + 'static, U: Unsigned + Data + Default, KH: Fn(&K) -> U + 'static, R: Data, RF: Fn(&K, &V1, &V2) -> R + 'static, LC: Lookup<K, Offset> + 'static, GC: Fn(u64) -> LC>(&self, stream2: &Stream<G, (D2, i32)>, kv1: F1, kv2: F2, part1: H1, part2: H2, key_h: KH, result: RF, look: &GC) -> Stream<G, (R, i32)> { ... }
}

Join implementations with parameterizable key selector functions.

Provided Methods

fn join_by_u<U: Unsigned + Data + Default, V1: Data + Default + 'static, V2: Data + Default + 'static, D2: Data, F1: Fn(D1) -> (U, V1) + 'static, F2: Fn(D2) -> (U, V2) + 'static, R: Data, RF: Fn(&U, &V1, &V2) -> R + 'static>(&self, other: &Stream<G, (D2, i32)>, kv1: F1, kv2: F2, result: RF) -> Stream<G, (R, i32)>

Matches elements of two streams using unsigned integers as the keys.

join_by_u takes a second input stream, two key-val selector functions, and a reduction function from an unsigned integer (the key) and two value references to the output type.

fn semijoin_by_u<U: Unsigned + Data + Default, V1: Data + Default + 'static, F1: Fn(D1) -> (U, V1) + 'static, RF: Fn(&U, &V1) -> D1 + 'static>(&self, other: &Stream<G, (U, i32)>, kv1: F1, result: RF) -> Stream<G, (D1, i32)>

Restricts the input stream to those elements whose unsigned integer key is present in the second stream.

semijoin_by_u takes a second stream, a key-val selector function, and a reconstruction function. Records are produced in the output if their key matches an unsigned integer present in the second stream. The key-val selector and reconstruction function are available to help avoid storing a redundant copy of the key in the value payload.

fn join_by<K: Data + 'static, V1: Data + 'static, V2: Data + 'static, D2: Data, F1: Fn(D1) -> (K, V1) + 'static, F2: Fn(D2) -> (K, V2) + 'static, U: Unsigned + Data + Default, KH: Fn(&K) -> U + 'static, R: Data, RF: Fn(&K, &V1, &V2) -> R + 'static>(&self, other: &Stream<G, (D2, i32)>, kv1: F1, kv2: F2, key_h: KH, result: RF) -> Stream<G, (R, i32)>

Matches elements of two streams using a key function.

fn semijoin_by<K: Data + 'static, V1: Data + 'static, F1: Fn(D1) -> (K, V1) + 'static, KH: Fn(&K) -> u64 + 'static, RF: Fn(&K, &V1) -> D1 + 'static>(&self, other: &Stream<G, (K, i32)>, kv1: F1, key_h: KH, result: RF) -> Stream<G, (D1, i32)>

Restricts the input stream to those elements whose key is present in the second stream.

fn join_by_core<K: Data, V1: Data, V2: Data, D2: Data, F1: Fn(D1) -> (K, V1) + 'static, F2: Fn(D2) -> (K, V2) + 'static, H1: Fn(&D1) -> u64 + 'static, H2: Fn(&D2) -> u64 + 'static, U: Unsigned + Data + Default, KH: Fn(&K) -> U + 'static, R: Data, RF: Fn(&K, &V1, &V2) -> R + 'static, LC: Lookup<K, Offset> + 'static, GC: Fn(u64) -> LC>(&self, stream2: &Stream<G, (D2, i32)>, kv1: F1, kv2: F2, part1: H1, part2: H2, key_h: KH, result: RF, look: &GC) -> Stream<G, (R, i32)>