Trait differential_dataflow::operators::threshold::Threshold

pub trait Threshold<G: Scope, K: Data> where G::Timestamp: Lattice {
    fn threshold<L>(&self, logic: L) -> Collection<G, K> where L: Fn(&K, Delta) -> Delta + 'static;
    fn threshold_u<L>(&self, logic: L) -> Collection<G, K> where K: Unsigned + Default, L: Fn(&K, Delta) -> Delta + 'static;

    fn distinct(&self) -> Collection<G, K> { ... }
    fn distinct_u(&self) -> Collection<G, K> where K: Unsigned + Default { ... }
}

Extension trait for the threshold differential dataflow method

Required Methods

fn threshold<L>(&self, logic: L) -> Collection<G, K> where L: Fn(&K, Delta) -> Delta + 'static

Groups records by their first field, and applies reduction logic to the associated values.

It would be nice for this to be generic over possible arrangements of data, but it seems that Rust ICEs when I try this. I'm not exactly sure how one would specify the arrangement (the type is used in logic, by way of its associated iterator types, but not clearly enough to drive type inference), but we don't have that problem yet anyhow.

In any case, it would be great if the same implementation could handle (K,V) pairs and K elements treated as (K, ()) pairs.

fn threshold_u<L>(&self, logic: L) -> Collection<G, K> where K: Unsigned + Default, L: Fn(&K, Delta) -> Delta + 'static

Groups records by their first field, when this field implements Unsigned.

This method uses a Vec<Option<_>> as its internal storage, allocating enough memory to directly index based on the first fields. This can be very useful when these fields are small integers, but it can be expensive if they are large and sparse.

Provided Methods

fn distinct(&self) -> Collection<G, K>

Collects distinct elements.

fn distinct_u(&self) -> Collection<G, K> where K: Unsigned + Default

Collects distinct elements, when they implement Unsigned.