Module differential_dataflow::trace::cursor

Traits and types for navigating order sequences of update tuples.

The Cursor trait contains several methods for efficiently navigating ordered collections of tuples of the form (key, val, time, diff). The cursor is different from an iterator both because it allows navigation on multiple levels (key and val), but also because it supports efficient seeking (via the seek_key and seek_val methods).

Modules

cursor_list
A generic cursor implementation merging multiple cursors.

Traits

Cursor
A cursor for navigating ordered (key, val, time, diff) updates.

CursorDebug
Debugging and testing utilities for Cursor.