Struct differential_dataflow::trace::cursor::cursor_list::CursorList

pub struct CursorList<K, V, T, R, C: Cursor<K, V, T, R>> { /* fields omitted */ }

Provides a cursor interface over a list of cursors.

The CursorList keeps its cursor ordered by key, and cursors with the smallest key are then ordered by val.

At any point, the current elements (keys, vals) are determined by the number of leading equivalent keys and equivalent values in self.cursors. Although they are implicit in self.cursors, We maintain these values explicitly, in self.equiv_keys and self.equiv_vals. We also track the number of valid keys and valid values, to avoid continually re-considering cursors in invalid states.

Methods

impl<K, V, T, R, C: Cursor<K, V, T, R>> CursorList<K, V, T, R, C> where K: Ord,
V: Ord

fn new(cursors: Vec<C>) -> Self

Creates a new cursor list from pre-existing cursors.