Struct differential_dataflow::trace::cursor::cursor_list_neu::CursorList

pub struct CursorList<K, V, T, R, C: Cursor<K, V, T, R>> { /* fields omitted */ }

Provides a cursor interface over a list of cursors.

The CursorList tracks the indices of cursors with the minimum key, and the the indices of cursors with the minimum key and minimum value. It performs no clever management of these sets otherwise.

Methods

impl<K, V, T, R, C: Cursor<K, V, T, R>> CursorList<K, V, T, R, C> where
K: Ord,
V: Ord,

pub fn new(cursors: Vec<C>, storage: &Vec<C::Storage>) -> Self

Creates a new cursor list from pre-existing cursors.