Struct differential_dataflow::trace::description::Description

pub struct Description<Time> { /* fields omitted */ }

Describes an interval of partially ordered times.

A Description indicates a set of partially ordered times, and a moment at which they are observed. The lower and upper frontiers bound the times contained within, and the since frontier indicates a moment at which the times were observed. If since is strictly in advance of lower, the contained times may be "advanced" to times which appear equivalent to any time after since.

Methods

impl<Time: Clone> Description<Time>

fn new(lower: &[Time], upper: &[Time], since: &[Time]) -> Self

Returns a new description from its component parts.

impl<Time> Description<Time>

fn lower(&self) -> &[Time]

The lower envelope for times in the interval.

fn upper(&self) -> &[Time]

The upper envelope for times in the interval.

fn since(&self) -> &[Time]

Times from whose future the interval may be observed.