Struct differential_dataflow::trace::implementations::spine::Spine

pub struct Spine<K, V, T: Lattice + Ord, R: Diff, B: Batch<K, V, T, R>> { /* fields omitted */ }

An append-only collection of update tuples.

A spine maintains a small number of immutable collections of update tuples, merging the collections when two have similar sizes. In this way, it allows the addition of more tuples, which may then be merged with other immutable collections.