Struct differential_dataflow::trace::implementations::spine_fueled::Spine

pub struct Spine<K, V, T: Lattice + Ord, R: Diff, B: Batch<K, V, T, R>> { /* fields omitted */ }

An append-only collection of update tuples.

A spine maintains a small number of immutable collections of update tuples, merging the collections when two have similar sizes. In this way, it allows the addition of more tuples, which may then be merged with other immutable collections.

Methods

impl<K, V, T, R, B> Spine<K, V, T, R, B> where
K: Ord + Clone,
V: Ord + Clone,
T: Lattice + Ord + Clone + Debug,
R: Diff,
B: Batch<K, V, T, R>,

pub fn with_effort(effort: usize) -> Self

Allocates a fueled Spine with a specified effort multiplier.

This trace will merge batches progressively, with each inserted batch applying a multiple of the batch's length in effort to each merge. The effort parameter is that multiplier. This value should be at least one for the merging to happen; a value of zero is not helpful.