Module differential_dataflow::trace::layers

Traits and types for building trie-based indices.

The trie structure has each each element of each layer indicate a range of elements in the next layer. Similarly, ranges of elements in the layer itself may correspond to single elements in the layer above.

Modules

hashed
Implementation using ordered keys with hashes and robin hood hashing.

ordered
Implementation using ordered keys and exponential search.

ordered_leaf
Implementation using ordered keys and exponential search.

unordered
Implementation using ordered keys and exponential search.

weighted
Implementation using ordered keys and exponential search.

Traits

Builder
A type used to assemble collections.

Cursor
A type supporting navigation.

MergeBuilder
A type used to assemble collections by merging other instances.

Trie
A collection of tuples, and types for building and enumerating them.

TupleBuilder
A type used to assemble collections from ordered sequences of tuples.