Struct differential_dataflow::trace::layers::ordered::OrderedCursor

pub struct OrderedCursor<K: Ord, L: Cursor> {
    pub child: L,
    // some fields omitted
}

A cursor with a child cursor that is updated as we move.

Fields
child: L

The cursor for the trie layer below this one.