Trait differential_dataflow::trace::layers::Cursor

pub trait Cursor<Storage> {
    type Key;
    fn key<'a>(&self, storage: &'a Storage) -> &'a Self::Key;

    fn step(&mut self, storage: &Storage);

    fn seek(&mut self, storage: &Storage, key: &Self::Key);

    fn valid(&self, storage: &Storage) -> bool;

    fn rewind(&mut self, storage: &Storage);

    fn reposition(&mut self, storage: &Storage, lower: usize, upper: usize);
}

A type supporting navigation.

The precise meaning of this navigation is not defined by the trait. It is likely that having navigated around, the cursor will be different in some other way, but the Cursor trait does not explain how this is so.

Associated Types

type Key

The type revealed by the cursor.

Required Methods

fn key<'a>(&self, storage: &'a Storage) -> &'a Self::Key

Reveals the current key.

fn step(&mut self, storage: &Storage)

Advances the cursor by one element.

fn seek(&mut self, storage: &Storage, key: &Self::Key)

Advances the cursor until the location where key would be expected.

fn valid(&self, storage: &Storage) -> bool

Returns true if the cursor points at valid data. Returns false if the cursor is exhausted.

fn rewind(&mut self, storage: &Storage)

Rewinds the cursor to its initial state.

fn reposition(&mut self, storage: &Storage, lower: usize, upper: usize)

Repositions the cursor to a different range of values.