Trait differential_dataflow::trace::layers::MergeBuilder

pub trait MergeBuilder: Builder {
    fn with_capacity(other1: &Self::Trie, other2: &Self::Trie) -> Self;
    fn copy_range(&mut self, other: &Self::Trie, lower: usize, upper: usize);
    fn push_merge(&mut self,
                  other1: (&Self::Trie, usize, usize),
                  other2: (&Self::Trie, usize, usize))
                  -> usize;
}

A type used to assemble collections by merging other instances.

Required Methods

fn with_capacity(other1: &Self::Trie, other2: &Self::Trie) -> Self

Allocates an instance of the builder with sufficient capacity to contain the merged data.

fn copy_range(&mut self, other: &Self::Trie, lower: usize, upper: usize)

Copies sub-collections of other into this collection.

fn push_merge(&mut self,
other1: (&Self::Trie, usize, usize),
other2: (&Self::Trie, usize, usize))
-> usize

Merges two sub-collections into one sub-collection.