Trait differential_dataflow::Data

pub trait Data: Data + Hash + Ord + Debug {
    fn hashed(&self) -> u64 { ... }
}

A composite trait for data types usable in differential dataflow.

Provided Methods

fn hashed(&self) -> u64

Extracts a u64 suitable for distributing and sorting the data.

The default implementation use FnvHasher. It might be nice to couple this more carefully with the implementor, to allow it to drive the distribution and sorting techniques. For example, dense unsigned integers would like a different function, but also must communicate that a HashMap is not the best way to use their values.