Trait differential_dataflow::TestScope

pub trait TestScope: Scope where Self::Timestamp: Lattice { }

An extension of timely's Scope trait requiring timestamps implement LeastUpperBound.

The intent of this trait is that it could be used as the constraint for collections, removing the need to put G::Timestamp: LeastUpperBound everywhere.