Struct differential_dataflow::difference::DiffPair

pub struct DiffPair<R1: Diff, R2: Diff> {
    pub element1: R1,
    pub element2: R2,
}

The difference defined by a pair of difference elements.

This type is essentially a "pair", though in Rust the tuple types do not derive the numeric traits we require, and so we need to emulate the types ourselves. In the interest of ergonomics, we may eventually replace the numeric traits with our own, so that we can implement them for tuples and allow users to ignore details like these.

Fields
element1: R1

The first element in the pair.

element2: R2

The second element in the pair.

Methods

impl<R1: Diff, R2: Diff> DiffPair<R1, R2>

pub fn new(elt1: R1, elt2: R2) -> Self

Creates a new Diff pair from two elements.