Trait differential_dataflow::lattice::Lattice

pub trait Lattice: PartialOrd {
    fn min() -> Self;
    fn max() -> Self;
    fn join(&self, &Self) -> Self;
    fn meet(&self, &Self) -> Self;

    fn advance_by(&self, frontier: &[Self]) -> Option<Self> where Self: Sized { ... }
}

A bounded partially ordered type supporting joins and meets.

Required Methods

fn min() -> Self

The smallest element of the type.

fn max() -> Self

The largest element of the type.

fn join(&self, &Self) -> Self

The smallest element greater than both arguments.

fn meet(&self, &Self) -> Self

The largest element lesser than both arguments.

Provided Methods

fn advance_by(&self, frontier: &[Self]) -> Option<Self> where Self: Sized

Advances self to the largest time indistinguishable under frontier.