Struct differential_dataflow::operators::arrange::Arranged

pub struct Arranged<G: Scope, T: Trace<Index=G::Timestamp>> where T::Key: Data, T::Value: Data, G::Timestamp: Lattice, &'a T: TraceRef<'a, T::Key, T::Index, T::Value> {
    pub stream: Stream<G, (Vec<T::Key>, Vec<u32>, Vec<(T::Value, i32)>)>,
    pub trace: Rc<RefCell<T>>,
}

A collection of (K,V) values as a timely stream and shared trace.

An Arranged performs the task of arranging a keyed collection once, allowing multiple differential operators to use the same trace. This saves on computation and memory, in exchange for some cognitive overhead in writing differential operators: each must pay enough care to signals from the stream field to know the subset of trace it has logically received.

Fields

stream

A stream containing arranged updates.

This stream may be reduced to just the keys, with the expectation that looking up the updated values in the associated trace should be more efficient than cloning the values. Users who need the values can obviously use the original stream.

trace

A shared trace, updated by the Arrange operator and readable by others.

Methods

impl<G: Scope, T: Trace<Index=G::Timestamp>> Arranged<G, T> where T::Key: Data, T::Value: Data, G::Timestamp: Lattice, &'a T: TraceRef<'a, T::Key, T::Index, T::Value>

fn as_collection(&self) -> Collection<G, (T::Key, T::Value)>

Flattens the stream into a Collection.

This operator is not obviously more efficient than using the Collection that is input to the Arrange operator. However, this may be the only way to recover a collection from only the Arranged type.