Struct differential_dataflow::operators::arrange::Arranged

pub struct Arranged<G: Scope, K, V, R, T> where
    G::Timestamp: Lattice + Ord,
    T: TraceReader<K, V, G::Timestamp, R> + Clone,  {
    pub stream: Stream<G, BatchWrapper<T::Batch>>,
    pub trace: T,
}

An arranged collection of (K,V) values.

An Arranged allows multiple differential operators to share the resources (communication, computation, memory) required to produce and maintain an indexed representation of a collection.

Fields
stream: Stream<G, BatchWrapper<T::Batch>>

A stream containing arranged updates.

This stream contains the same batches of updates the trace itself accepts, so there should be no additional overhead to receiving these records. The batches can be navigated just as the batches in the trace, by key and by value.

trace: T

A shared trace, updated by the Arrange operator and readable by others.

Methods

impl<G: Scope, K, V, R, T> Arranged<G, K, V, R, T> where
G::Timestamp: Lattice + Ord,
T: TraceReader<K, V, G::Timestamp, R> + Clone,

pub fn enter<'a, TInner>(
&self,
child: &Child<'a, G, TInner>
) -> Arranged<Child<'a, G, TInner>, K, V, R, TraceEnter<K, V, G::Timestamp, R, T, TInner>> where
T::Batch: Clone,
K: 'static,
V: 'static,
G::Timestamp: Clone + Default + 'static,
TInner: Lattice + Timestamp + Clone + Default + 'static,
R: 'static,

Brings an arranged collection into a nested scope.

This method produces a proxy trace handle that uses the same backing data, but acts as if the timestamps have all been extended with an additional coordinate with the default value. The resulting collection does not vary with the new timestamp coordinate.

pub fn as_collection<D: Data, L>(&self, logic: L) -> Collection<G, D, R> where
R: Diff,
T::Batch: Clone + 'static,
K: Clone,
V: Clone,
L: Fn(&K, &V) -> D + 'static,

Flattens the stream into a Collection.

The underlying Stream<G, BatchWrapper<T::Batch>> is a much more efficient way to access the data, and this method should only be used when the data need to be transformed or exchanged, rather than supplied as arguments to an operator using the same key-value structure.

pub fn flat_map_ref<I, L>(&self, logic: L) -> Collection<G, I::Item, R> where
R: Diff,
T::Batch: Clone + 'static,
K: Clone,
V: Clone,
I: IntoIterator,
I::Item: Data,
L: Fn(&K, &V) -> I + 'static,

Extracts elements from an arrangement as a collection.

The supplied logic may produce an iterator over output values, allowing either filtering or flat mapping as part of the extraction.