Trait differential_dataflow::operators::arrange::ArrangeByKey

pub trait ArrangeByKey<G: Scope, K: Data, V: Data> where G::Timestamp: Lattice {
    fn arrange_by_key<U: Unsigned + Default, KH: Fn(&K) -> U + 'static, Look: Lookup<K, Offset> + 'static, LookG: Fn(u64) -> Look>(&self, key_h: KH, look: LookG) -> Arranged<G, BasicTrace<K, G::Timestamp, V, Look>>;
}

Arranges something as (Key,Val) pairs.

Required Methods

fn arrange_by_key<U: Unsigned + Default, KH: Fn(&K) -> U + 'static, Look: Lookup<K, Offset> + 'static, LookG: Fn(u64) -> Look>(&self, key_h: KH, look: LookG) -> Arranged<G, BasicTrace<K, G::Timestamp, V, Look>>

Arranges a stream of (Key, Val) updates by Key.

This operator arranges a stream of values into a shared trace, whose contents it maintains. This trace is current for all times completed by the output stream, which can be used to safely identify the stable times and values in the trace.