Trait differential_dataflow::operators::count::CountTotalCore

pub trait CountTotalCore<G: Scope, K: Data, R: Diff> where
    G::Timestamp: TotalOrder + Lattice + Ord, {
    fn count_total_core(&self) -> Collection<G, (K, R), isize>;
}

Extension trait for the group_arranged differential dataflow method.

Required Methods

fn count_total_core(&self) -> Collection<G, (K, R), isize>

Applies group to arranged data, and returns an arrangement of output data.

This method is used by the more ergonomic group, distinct, and count methods, although it can be very useful if one needs to manually attach and re-use existing arranged collections.

Examples

extern crate timely;
extern crate differential_dataflow;

use differential_dataflow::input::Input;
use differential_dataflow::operators::arrange::Arrange;
use differential_dataflow::operators::count::CountTotalCore;
use differential_dataflow::trace::Trace;
use differential_dataflow::trace::implementations::ord::OrdKeySpine;
use differential_dataflow::hashable::OrdWrapper;

fn main() {
    ::timely::example(|scope| {

        // wrap and order input, then group manually.
        scope.new_collection_from(1 .. 10u32).1
             .map(|x| (OrdWrapper { item: x / 3 }, ()))
             .arrange(OrdKeySpine::new())
             .count_total_core();
    });
}