Module differential_dataflow::operators::iterate

Iterative application of a differential dataflow fragment.

The iterate operator takes as an argument a closure from a differential dataflow stream to a stream of the same type. The output are differences which accumulate to the result of applying this closure a specified number of times.

The implementation of iterate does not directly apply the closure, but rather establishes an iterative timely dataflow subcomputation, in which differences circulate until they dissipate (indicating that the computation has reached fixed point), or until some number of iterations have passed.

Examples

The example repeatedly divides even numbers by two, and leaves odd numbers as they are. Although some numbers may take multiple iterations to converge, converged numbers have no overhead in subsequent iterations.

// repeatedly divide out factors of two.
let limits = numbers.iterate(|values| {
    values.map(|(x,w) if x % 2 == 0 { (x / 2, w) } else { (x, w) })
});

If anyone knows of a form of the Collatz conjecture in which the iterates achieve fixed point, rather than cycling, let me know!

Traits

IterateExt
An extension trait for the iterate method.