Struct differential_dataflow::operators::iterate::Variable

pub struct Variable<'a, G: Scope, D: Data> where G::Timestamp: Lattice {
    // some fields omitted
}

A differential dataflow collection variable

The Variable struct allows differential dataflow programs requiring more sophisticated iterative patterns than singly recursive iteration. For example: in mutual recursion two collections evolve simultaneously.

Methods

impl<'a, G: Scope, D: Data> Variable<'a, G, D> where G::Timestamp: Lattice

fn from(source: Collection<Child<'a, G, u64>, D>) -> Variable<'a, G, D>

Creates a new Variable and a Stream representing its output, from a supplied source stream.

fn set(self, result: &Collection<Child<'a, G, u64>, D>) -> Collection<Child<'a, G, u64>, D>

Adds a new source of data to the Variable.