Trait differential_dataflow::operators::join::Join

pub trait Join<G: Scope, K: Data, V: Data>: JoinBy<G, (K, V)> where G::Timestamp: LeastUpperBound {
    fn join<V2: Data>(&self, other: &Collection<G, (K, V2)>) -> Collection<G, (K, V, V2)> { ... }
    fn join_map<V2: Data, D: Data, R>(&self, other: &Collection<G, (K, V2)>, logic: R) -> Collection<G, D> where R: Fn(&K, &V, &V2) -> D + 'static { ... }
}

Join implementations for (key,val) data.

The Join trait provides default implementations of join for streams of data with

Provided Methods

fn join<V2: Data>(&self, other: &Collection<G, (K, V2)>) -> Collection<G, (K, V, V2)>

Matches pairs (key,val1) and (key,val2) based on key.

The join method requires that the two collections both be over pairs of records, and the first element of the pair must be of the same type. Given two such collections, each pair of records (key,val1) and (key,val2) with a matching key produces a (key, val1, val2) output record.

Examples

extern crate timely;
use timely::dataflow::operators::{ToStream, Inspect};
use differential_dataflow::operators::Join;

timely::example(|scope| {
    let col1 = vec![((0,0),1),((1,2),1)].into_iter().to_stream(scope);
    let col2 = vec![((0,'a'),1),((1,'B'),1)].into_iter().to_stream(scope);

    // should produce triples `(0,0,'a')` and `(1,2,'B')`.
    col1.join(&col2).inspect(|x| println!("observed: {:?}", x));
});

fn join_map<V2: Data, D: Data, R>(&self, other: &Collection<G, (K, V2)>, logic: R) -> Collection<G, D> where R: Fn(&K, &V, &V2) -> D + 'static

Matches pairs of (key,val1) and (key,val2) records based on key and applies a reduction function.