Module differential_dataflow::trace

Traits and datastructures representing a collection trace.

A collection trace is a set of updates of the form (key, val, time, diff), which determine the contents of a collection at given times by accumulating updates whose time field is less or equal to the target field.

The Trace trait describes those types and methods that a data structure must implement to be viewed as a collection trace. This trait allows operator implementations to be generic with respect to the type of trace, and allows various data structures to be interpretable as multiple different types of trace.

Re-exports

pub use self::cursor::Cursor;

pub use self::description::Description;

Modules

abomonated_blanket_impls
Blanket implementations for reference counted batches.

cursor
Traits and types for navigating order sequences of update tuples.

description
Descriptions of intervals of partially ordered times.

implementations
Implementations of Trace and associated traits.

layers
Traits and types for building trie-based indices.

rc_blanket_impls
Blanket implementations for reference counted batches.

wrappers
Wrappers around trace implementations, providing derived views of updates.

Traits

Batch
An immutable collection of updates.

BatchReader
A batch of updates whose contents may be read.

Batcher
Functionality for collecting and batching updates.

Builder
Functionality for building batches from ordered update sequences.

Merger
Represents a merge in progress.

Trace
An append-only collection of (key, val, time, diff) tuples.

TraceReader
A trace whose contents may be read.

Functions

consolidate
Scans vec[off..] and consolidates differences of adjacent equivalent elements.

consolidate_by
Scans vec[off..] and consolidates differences of adjacent equivalent elements.