Trait differential_dataflow::trace::Batch

pub trait Batch<K, V, T, R>: BatchReader<K, V, T, R> where
    Self: Sized, {
    type Batcher: Batcher<K, V, T, R, Self>;
    type Builder: Builder<K, V, T, R, Self>;
    type Merger: Merger<K, V, T, R, Self>;
    fn begin_merge(&self, other: &Self) -> Self::Merger { ... }
}

An immutable collection of updates.

Associated Types

type Batcher: Batcher<K, V, T, R, Self>

A type used to assemble batches from disordered updates.

type Builder: Builder<K, V, T, R, Self>

A type used to assemble batches from ordered update sequences.

type Merger: Merger<K, V, T, R, Self>

A type used to progressively merge batches.

Provided Methods

fn begin_merge(&self, other: &Self) -> Self::Merger

Initiates the merging of consecutive batches.

The result of this method can be exercised to eventually produce the same result that a call to self.merge(other) would produce, but it can be done in a measured fashion. This can help to avoid latency spikes where a large merge needs to happen.