Trait differential_dataflow::trace::BatchReader

pub trait BatchReader<K, V, T, R> where
    Self: Sized, {
    type Cursor: Cursor<K, V, T, R, Storage = Self>;
    fn cursor(&self) -> Self::Cursor;

    fn len(&self) -> usize;

    fn description(&self) -> &Description<T>;

    fn lower(&self) -> &[T] { ... }

    fn upper(&self) -> &[T] { ... }
}

A batch of updates whose contents may be read.

This is a restricted interface to batches of updates, which support the reading of the batch's contents, but do not expose ways to construct the batches. This trait is appropriate for views of the batch, and is especially useful for views derived from other sources in ways that prevent the construction of batches from the type of data in the view (for example, filtered views, or views with extended time coordinates).

Associated Types

type Cursor: Cursor<K, V, T, R, Storage = Self>

The type used to enumerate the batch's contents.

Required Methods

fn cursor(&self) -> Self::Cursor

Acquires a cursor to the batch's contents.

fn len(&self) -> usize

The number of updates in the batch.

fn description(&self) -> &Description<T>

Describes the times of the updates in the batch.

Provided Methods

fn lower(&self) -> &[T]

All times in the batch are greater or equal to an element of lower.

fn upper(&self) -> &[T]

All times in the batch are not greater or equal to any element of upper.