Trait differential_dataflow::trace::Batcher

pub trait Batcher<K, V, T, R, Output: Batch<K, V, T, R>> {
    fn new() -> Self;
    fn push_batch(&mut self, batch: &mut Vec<((K, V), T, R)>);
    fn seal(&mut self, upper: &[T]) -> Output;
    fn frontier(&mut self) -> &[T];
}

Functionality for collecting and batching updates.

Required Methods

fn new() -> Self

Allocates a new empty batcher.

fn push_batch(&mut self, batch: &mut Vec<((K, V), T, R)>)

Adds an unordered batch of elements to the batcher.

fn seal(&mut self, upper: &[T]) -> Output

Returns all updates not greater or equal to an element of upper.

fn frontier(&mut self) -> &[T]

Returns the lower envelope of contained update times.