Trait differential_dataflow::trace::TraceReader

pub trait TraceReader<Key, Val, Time, R> {
    type Batch: BatchReader<Key, Val, Time, R> + Clone + 'static;
    type Cursor: Cursor<Key, Val, Time, R>;
    fn cursor_through(
        &mut self, 
        upper: &[Time]
    ) -> Option<(Self::Cursor, <Self::Cursor as Cursor<Key, Val, Time, R>>::Storage)>;

    fn advance_by(&mut self, frontier: &[Time]);

    fn advance_frontier(&mut self) -> &[Time];

    fn distinguish_since(&mut self, frontier: &[Time]);

    fn distinguish_frontier(&mut self) -> &[Time];

    fn map_batches<F: FnMut(&Self::Batch)>(&mut self, f: F);

    fn cursor(
        &mut self
    ) -> (Self::Cursor, <Self::Cursor as Cursor<Key, Val, Time, R>>::Storage) { ... }
}

A trace whose contents may be read.

This is a restricted interface to the more general Trace trait, which extends this trait with further methods to update the contents of the trace. These methods are used to examine the contents, and to update the reader's capabilities (which may release restrictions on the mutations to the underlying trace and cause work to happen).

Associated Types

type Batch: BatchReader<Key, Val, Time, R> + Clone + 'static

The type of an immutable collection of updates.

type Cursor: Cursor<Key, Val, Time, R>

The type used to enumerate the collections contents.

Required Methods

fn cursor_through(
&mut self,
upper: &[Time]
) -> Option<(Self::Cursor, <Self::Cursor as Cursor<Key, Val, Time, R>>::Storage)>

Acquires a cursor to the restriction of the collection's contents to updates at times not greater or equal to an element of upper.

This method is expected to work if called with an upper that (i) was an observed bound in batches from the trace, and (ii) the trace has not been advanced beyond upper. Practically, the implementation should be expected to look for a "clean cut" using upper, and if it finds such a cut can return a cursor. This should allow upper such as &[] as used by self.cursor(), though it is difficult to imagine other uses.

fn advance_by(&mut self, frontier: &[Time])

Advances the frontier of times the collection must be correctly accumulable through.

Practically, this allows the trace to advance times in updates it maintains as long as the advanced times still compare equivalently to any times greater or equal to some element of frontier. Times not greater or equal to some element of frontier may no longer correctly accumulate, so do not advance a trace unless you are quite sure you no longer require the distinction.

fn advance_frontier(&mut self) -> &[Time]

Reports the frontier from which all time comparisions should be accurate.

Times that are not greater or equal to some element of the advance frontier may accumulate inaccurately as the trace may have lost the ability to distinguish between such times. Accumulations are only guaranteed to be accurate from the frontier onwards.

fn distinguish_since(&mut self, frontier: &[Time])

Advances the frontier that may be used in cursor_through.

Practically, this allows the trace to merge batches whose upper frontier comes before frontier. The trace is likely to be annoyed or confused if you use a frontier other than one observed as an upper bound of an actual batch. This doesn't seem likely to be a problem, but get in touch if it is.

Calling distinguish_since(&[]) indicates that all batches may be merged at any point, which essentially disables the use of cursor_through with any parameter other than &[], which is the behavior of cursor.

fn distinguish_frontier(&mut self) -> &[Time]

Reports the frontier from which the collection may be subsetted.

The semantics are less elegant here, but the underlying trace will not merge batches in advance of this frontier, which ensures that operators can extract the subset of the trace at batch boundaries from this frontier onward. These boundaries may be used in cursor_through, whereas boundaries not in advance of this frontier are not guaranteed to return a cursor.

fn map_batches<F: FnMut(&Self::Batch)>(&mut self, f: F)

Maps some logic across the batches the collection manages.

This is currently used only to extract historical data to prime late-starting operators who want to reproduce the stream of batches moving past the trace. It could also be a fine basis for a default implementation of the cursor methods, as they (by default) just move through batches accumulating cursors into a cursor list.

Provided Methods

fn cursor(
&mut self
) -> (Self::Cursor, <Self::Cursor as Cursor<Key, Val, Time, R>>::Storage)

Provides a cursor over updates contained in the trace.