Struct differential_dataflow::trace::wrappers::rc::TraceRc

pub struct TraceRc<K, V, T, R, Tr> where
    T: Lattice + Ord + Clone + 'static,
    Tr: TraceReader<K, V, T, R>,  {
    pub wrapper: Rc<RefCell<TraceBox<K, V, T, R, Tr>>>,
    // some fields omitted
}

A handle to a shared trace.

As long as the handle exists, the wrapped trace should continue to exist and will not advance its timestamps past the frontier maintained by the handle. The intent is that such a handle appears as if it is a privately maintained trace, despite being backed by shared resources.

Fields
wrapper: Rc<RefCell<TraceBox<K, V, T, R, Tr>>>

Wrapped trace. Please be gentle when using.

Methods

impl<K, V, T, R, Tr> TraceRc<K, V, T, R, Tr> where
T: Lattice + Ord + Clone + 'static,
Tr: TraceReader<K, V, T, R>,

pub fn make_from(trace: Tr) -> (Self, Rc<RefCell<TraceBox<K, V, T, R, Tr>>>)

Allocates a new handle from an existing wrapped wrapper.